refactor(location): move packages
