--- ip/module.go
@@ -1,12 +1,0 @@
-package ip
-
-import (
-	"go.uber.org/fx"
-)
-
-var (
-	// Module for fx.
-	Module = fx.Options(
-		fx.Provide(NewDB),
-	)
-)
