+++ location/config.go
@@ -1,0 +1,10 @@
+package location
+
+import (
+	"github.com/alexfalkowski/standort/location/ip"
+)
+
+// Config for location.
+type Config struct {
+	IP ip.Config `yaml:"ip"`
+}
